requirements.txt
numpy
pandas
transformers
faiss-gpu
pillow
tqdm
jsonschema
click